Fix timeout in concurrency test
The Runner was calling init-store for each creation, which was taking
way longer after the init store changes.
Instead init-store is called once before the tests, and all creations
will skip it.

[#142588813]

+++ grootfs/integration/runner/runner.go
@@ -118,8 +118,11 @@ func (r Runner) runCmd(cmd *exec.Cmd) error {
 		return runErr
 
 	case <-time.After(r.Timeout):
-		pid := cmd.Process.Pid
-		syscall.Kill(pid, syscall.SIGKILL)
+		if process := cmd.Process; process != nil {
+			pid := process.Pid
+			syscall.Kill(pid, syscall.SIGKILL)
+		}
+
 		return errors.New(
 			fmt.Sprintf("command took more than %f seconds to finish", r.Timeout.Seconds()),
 		)
